WIP: Refactoring files to enable parallel nonblocking and mulithreaded development.
