+++ Makefile
@@ -7,8 +7,13 @@ else
     CFLAGS += -O2
 endif
 
-mkmimo: mkmimo.o
+ifdef MULTITHREADED
+mkmimo: mkmimo_multithreaded.o queue.o
 	$(CC) -o $@ $(LDFLAGS) $^
+else
+mkmimo: main.o utils.o mkmimo_nonblocking.o
+	$(CC) -o $@ $(LDFLAGS) $^
+endif
 
 PATH := $(shell pwd):$(PATH)
 export PATH
