+++ mkmimo_multithreaded.h
@@ -1,0 +1,8 @@
+#ifndef MKMIMO_MULTI_H
+#define MKMIMO_MULTI_H
+
+#include "mkmimo.h"
+
+int mkmimo_multithreaded(Input *inputs, Output *outputs);
+
+#endif /* MKMIMO_MULTI_H */
