+++ mkmimo_nonblocking.h
@@ -1,0 +1,8 @@
+#ifndef MKMIMO_NB_H
+#define MKMIMO_NB_H
+
+#include "mkmimo.h"
+
+int mkmimo_nonblocking(Inputs *inputs, Outputs *outputs);
+
+#endif /* MKMIMO_NB_H */
