+++ queue.c
@@ -4,15 +4,15 @@ typedef struct Node Node;
 typedef struct Queue Queue;
 
 Queue *new_queue() {
-  Queue *q = (Queue *) malloc(sizeof(Queue));
+  Queue *q = (Queue *)malloc(sizeof(Queue));
   q->first = NULL;
   q->last = NULL;
   return q;
 }
 
-void queue(Queue *q, int *elem) {
+void queue(Queue *q, void *elem) {
   // Create a new queue node
-  Node *new_node = (Node *) malloc(sizeof(Node));
+  Node *new_node = (Node *)malloc(sizeof(Node));
   new_node->elem = elem;
   new_node->next = NULL;
 
@@ -20,17 +20,15 @@ void queue(Queue *q, int *elem) {
   if (is_empty(q)) {
     q->first = new_node;
     q->last = new_node;
-  } else {           
+  } else {
     q->last->next = new_node;
     q->last = new_node;
   }
 }
 
-Node* peek(Queue *q) {
-  return q->first;
-}
+Node *peek(Queue *q) { return q->first; }
 
-int *dequeue(Queue *q) {   
+int *dequeue(Queue *q) {
   int *elem = q->first->elem;
   Node *next = q->first->next;
   free(q->first);
@@ -39,7 +37,8 @@ int *dequeue(Queue *q) {
   return elem;
 }
 
-bool is_empty(Queue *q) {       
-  if (q->first == NULL) return true;    
+bool is_empty(Queue *q) {
+  if (q->first == NULL)
+    return true;
   return false;
 }
